Improve reverb

+++ TOMB4/specific/dxsound.h
@@ -23,6 +23,6 @@ void DXFreeSounds();
 long S_SoundSampleIsPlaying(long num);
 void S_SoundSetPanAndVolume(long num, short pan, ushort volume);
 void S_SoundSetPitch(long num, long pitch);
-void S_SetReverbType(int reverb);
+void S_SetReverbType(long reverb);
 
 extern char* samples_buffer;
